Move new fluentd config map code to common controllers package
